Might work, might blow up. We'll find out.

+++ NewDot/NSURL+QueryStringConstructor.m
@@ -15,15 +15,26 @@
     return [[[NSURL alloc] fs_initWithString:URLString relativeToURL:baseURL queryParameters:params] autorelease];
 }
 
++ (id)fs_URLWithString:(NSString *)URLString relativeToURL:(NSURL *)baseURL queryParameters:(NSDictionary *)params tailParams:(NSString*)tailParams
+{
+    return [[[NSURL alloc] fs_initWithString:URLString relativeToURL:baseURL queryParameters:params tailParams:tailParams] autorelease];
+}
+
 - (id)fs_initWithString:(NSString*)URLString relativeToURL:(NSURL*)baseURL queryParameters:(NSDictionary*)params
+{
+    return [self fs_initWithString:URLString relativeToURL:baseURL queryParameters:params tailParams:nil];
+}
+
+- (id)fs_initWithString:(NSString *)URLString relativeToURL:(NSURL *)baseURL queryParameters:(NSDictionary *)params tailParams:(NSString*)tailParams
 {
     NSMutableString* _URLString = [[NSMutableString alloc] initWithString:URLString];
     if (params&&[params count]>0) {
         [_URLString appendString:@"?"];
         for (id key in [params allKeys])
             [_URLString appendFormat:@"%@=%@&", [[key description] stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding], [[[params objectForKey:key] description] stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding]];
-        [_URLString deleteCharactersInRange:NSMakeRange([_URLString length]-1, 1)]; // kill the trailing '&'
+        if (!tailParams) [_URLString deleteCharactersInRange:NSMakeRange([_URLString length]-1, 1)]; // kill the trailing '&'
     }
+    [_URLString appendString:tailParams];
     
     return [self initWithString:[_URLString autorelease] relativeToURL:baseURL];
 }
